Fix explicit variable referencing order for isset in filters
Unlikely we'll hit this, but the tests are broken because the
`isset` looks like it doesn't work properly with a chain of
objects using offsetExists (in the complex setup d1 has) and
returns false prematurely.

Dereferencing the first record for the relation appears then
to cause the isset to operate on the last element properly.